# portfolio_quant_site/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Quantitative Portfolio Management System (QPMS) API
  version: 1.0.0
  description: >
    REST API for the Quantitative Portfolio Management System.
    Covers users, stocks (Yahoo-enriched master data), portfolios, holdings,
    stock prices, transactions, portfolio analytics, factor sorting and
    asset-allocation optimizers. Money values use fixed-precision decimals
    (never floats). Single-user, no authentication.
  contact:
    name: QPMS
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/v1/stocks:
    get:
      tags: [Stocks]
      summary: List known stocks (name and/or symbol)
      operationId: listStocks
      parameters:
        - name: search
          in: query
          description: Filter by symbol or name substring.
          schema:
            type: string
        - name: sector
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of stocks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockSummary'
  /api/v1/stocks/{stockId}:
    get:
      tags: [Stocks]
      summary: Get full stock detail (sector, summary, market cap, ...)
      operationId: getStock
      parameters:
        - $ref: '#/components/parameters/StockId'
      responses:
        '200':
          description: Stock detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stock'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/stocks/{stockId}/prices:
    get:
      tags: [Stocks]
      summary: Price history (OHLC candles) for a stock
      description: >
        Returns candles for the requested interval, optionally bounded by a
        time frame. Used to render the price chart and to compute risk,
        volatility, returns, Sharpe ratio and drawdown.
      operationId: getStockPrices
      parameters:
        - $ref: '#/components/parameters/StockId'
        - name: interval
          in: query
          description: Candle interval (1d, 1h, 5m, ...).
          schema:
            type: string
            default: 1d
        - name: start
          in: query
          description: Inclusive start timestamp (ISO 8601).
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          description: Inclusive end timestamp (ISO 8601).
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Candle series.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PriceCandle'
  /api/v1/stocks/{stockId}/quote:
    get:
      tags: [Stocks]
      summary: Latest live price for a stock
      operationId: getStockQuote
      parameters:
        - $ref: '#/components/parameters/StockId'
      responses:
        '200':
          description: Latest quote.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # PORTFOLIOS
  # ============================================================
...
components:
  parameters:
...
    StockId:
      name: stockId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 500
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
  responses:
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
      required: [detail]

    # ---------- Core entities ----------
...
    StockSummary:
      type: object
      properties:
        stock_id:
          type: integer
          format: int64
        symbol:
          type: string
        short_name:
          type: string
          nullable: true
        sector:
          type: string
          nullable: true
      required: [stock_id, symbol]
    Stock:
      allOf:
        - $ref: '#/components/schemas/StockSummary'
        - type: object
          properties:
            exchange:
              type: string
              nullable: true
            quote_type:
              type: string
              nullable: true
            long_name:
              type: string
              nullable: true
            currency:
              type: string
              nullable: true
            country:
              type: string
              nullable: true
            industry:
              type: string
              nullable: true
            website:
              type: string
              nullable: true
            business_summary:
              type: string
              nullable: true
            market_cap:
              type: integer
              format: int64
              nullable: true
            shares_outstanding:
              type: integer
              format: int64
              nullable: true
            first_seen_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    PriceCandle:
      type: object
      properties:
        ts:
          type: string
          format: date-time
        interval:
          type: string
        open:
          type: string
        high:
          type: string
        low:
          type: string
        close:
          type: string
        adj_close:
          type: string
          nullable: true
        volume:
          type: integer
          format: int64
          nullable: true
        dividend:
          type: string
          nullable: true
        stock_split:
          type: string
          nullable: true
      required: [ts, interval, open, high, low, close]
    Quote:
      type: object
      properties:
        stock_id:
          type: integer
          format: int64
        symbol:
          type: string
        price:
          type: string
        ts:
          type: string
          format: date-time
      required: [stock_id, symbol, price, ts]
...